fix tests and documentation
