+++ model_compression_toolkit/core/common/quantization/debug_config.py
@@ -29,7 +29,7 @@ class DebugConfig:
          enabled) or not. Can be used to pinpoint problematic layers in the quantization process.
         network_editor (List[EditRule]): A list of rules and actions to edit the network for quantization.
         simulate_scheduler (bool): Simulate scheduler behavior to compute operators' order and cuts.
-        bypass (bool): A flag to enable MCT bypass.
+        bypass (bool): A flag to enable MCT bypass, which skips MCT runner and returns the input model unchanged.
     """
 
     analyze_similarity: bool = False
